Static Code Analysis Clean-Up, String [Cheap] Replace Method Tweaks
- Added a string extension that is effectively string.Replace() with a non-Ordinal StringComparison. Removed the "case-insensitive" option for CheapReplace() (this option didn't actually do exactly what it was intended to do anyway) and replaced it with an optional StringComparison argument that can then be used to actually achieve what the old "case-insensitive" option was supposed to do.
- Switched the CheapReplace() calls in the telemetry methods to regular Replace() calls, albeit ones that use the new, non-Ordinal StringComparison-enabling extension. Also added extra info to the CheapReplace methods' documentation to explicitly mention that they are meant to be used only when the new value is obtained through a non-trivial method call or function, otherwise CheapReplace is actually slower than a regular Replace (since you get overhead from doing a Contains check one more time than necessary and also for doing a function call to get your new value, but get no performance benefit for if the existing string does not contain anything that would need to be replaced).
- Fixed an extremely minor/niche case scenario where the number for how much ammo a weapon has currently available ignored the character sheet language option for number formatting and instead used the user system's number formatting options. This is such a minor fix that I don't think it's worth a changelog entry.

+++ Chummer/UI/Chummy.cs
@@ -165,8 +165,8 @@ namespace Chummer
 
             // Draw an ellipse 1/2 the size of the eye
             // centered at (px, py).
-            int x = (int)(px - wid / 4);
-            int y = (int)(py - hgt / 4);
+            int x = (int)(px - (double)wid / 4);
+            int y = (int)(py - (double)hgt / 4);
             int width = wid / 2;
             int height = hgt / 2;
             try
